DFI-1820 architectural changes
