+++ app/uk/gov/hmrc/disareturnstestsupportapi/controllers/GenerateReportController.scala
@@ -23,7 +23,7 @@ import uk.gov.hmrc.disareturnstestsupportapi.config.AppConfig
 import uk.gov.hmrc.disareturnstestsupportapi.controllers.actions.AuthAction
 import uk.gov.hmrc.disareturnstestsupportapi.controllers.parsers.StrictJsonBodyParser
 import uk.gov.hmrc.disareturnstestsupportapi.models.GenerateReportRequest
-import uk.gov.hmrc.disareturnstestsupportapi.models.common._
+import uk.gov.hmrc.disareturnstestsupportapi.models.validators._
 import uk.gov.hmrc.disareturnstestsupportapi.models.errors._
 import uk.gov.hmrc.disareturnstestsupportapi.service.GenerateReportService
 import uk.gov.hmrc.disareturnstestsupportapi.utils.RequestParser
