+++ app/uk/gov/hmrc/disareturnstestsupportapi/models/validators/MonthValidator.scala
@@ -14,13 +14,25 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.disareturnstestsupportapi.models.common
+package uk.gov.hmrc.disareturnstestsupportapi.models.validators
 
-import scala.util.matching.Regex
+object MonthValidator {
+  private val validMonths: Set[String] = Set(
+    "JAN",
+    "FEB",
+    "MAR",
+    "APR",
+    "MAY",
+    "JUN",
+    "JUL",
+    "AUG",
+    "SEP",
+    "OCT",
+    "NOV",
+    "DEC"
+  )
 
-object IsaRefValidator {
-  private val isaRefRegex: Regex = "^[z|Z][0-9]{4}$".r
+  def isValid(month: String): Boolean =
+    Option(month).exists(m => validMonths.contains(m.toUpperCase))
 
-  def isValid(ref: String): Boolean =
-    isaRefRegex.pattern.matcher(ref).matches()
 }
